pyproject.toml
[project]
name = "vidagent-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
]
